refactor: 스타일 코드와 프로덕션 코드 분리 및 장바구니 컴포넌트 공용화 제거
