+++ src/components/CartItem/CartItemContent.tsx
@@ -1,0 +1,23 @@
+import CountButtonContainer from './CountButtonContainer/CountButtonContainer';
+import { image, itemBody, itemWrapper, price } from '../styles/item';
+
+import { CartItemProps } from '@/types/cartItem';
+
+interface CartItemContentProps {
+  item: CartItemProps;
+}
+
+const CartItemContent = ({ item }: CartItemContentProps) => {
+  return (
+    <div css={itemBody}>
+      <img css={image} src={item.product.imageUrl} width={112} height={112} />
+      <div css={itemWrapper}>
+        <span>{item.product.name}</span>
+        <span css={price}>{item.product.price.toLocaleString('ko-KR')}원</span>
+        <CountButtonContainer item={item} />
+      </div>
+    </div>
+  );
+};
+
+export default CartItemContent;
